requirements.txt
torch
torchvision
timm
matplotlib
Pillow
